created TempBucket with auto-cleaning upon init

+++ server/buckets.js
@@ -1,3 +1,6 @@
+const fs = require('fs');
+const uploadsDir = `${__dirname}/uploads/`;
+
 class Bucket{
     constructor(name, mountPoint, type) {
         this.name = name;
@@ -13,40 +16,59 @@ class StandardBucket extends Bucket {
         super(name, mountPoint, 'std');
     }
 }
-// class TempBucket extends Bucket {
-//     constructor(name, mountPoint, timeMin) {
-//         super(name, mountPoint, 'tmp');
-//         this.timeMin = timeMin;
-//         this.cleaner = setInterval(() => {
-//             rmDir(`${uploadsDir}${this.mountPoint}`, false);
-//         }, timeMin * 1000 * 60);
-//     }
-// }
+class TempBucket extends Bucket {
+    constructor(name, mountPoint, timeMin) {
+        super(name, mountPoint, 'tmp');
+        this.timeMin = timeMin; //time after which files are deleted
+        this.maxError = 0.1; //max percent of extra time files might live before deletion (e.g. files may live 0.1 ==> 10% longer than timeMin)
+        this.cleaner = setInterval(() => {
+            console.log('running cleaner for timeMin: ', this.timeMin);
+            rmDirTimed(`${uploadsDir}${this.mountPoint}`, false, this.timeMin);
+        }, (timeMin * 1000 * 60) * this.maxError);
+    }
+}
 
 function initBuckets() {
-    // const numTempBuckets = 3; //number of temp buckets
-    // const tempBucketMultiplier = 2; //multiplier between temp buckets' cleaning intervals
-    // const tempBucketInit = 1; //fastest cleaning temp bucket's interval (min)
+    const numTempBuckets = 3; //number of temp buckets
+    const tempBucketMultiplier = 2; //multiplier between temp buckets' cleaning intervals
+    const tempBucketInit = 1; //fastest cleaning temp bucket's interval (min)
     const buckets = {};
     buckets['std'] = new StandardBucket('std', 'std/');
-    // for(let i = tempBucketInit; i < Math.pow(tempBucketMultiplier, numTempBuckets); i *= tempBucketMultiplier) {
-    //     buckets.push(new TempBucket(`tmp${i}`, `/tmp${i}/`, i)));  
-    // }
+    for(let i = tempBucketInit; i < tempBucketInit * Math.pow(tempBucketMultiplier, numTempBuckets); i *= tempBucketMultiplier) {
+        buckets[`tmp${i}`] = new TempBucket(`tmp${i}`, `tmp${i}/`, i);  
+    }
     return buckets;
 }
-function rmDir(dirPath, removeSelf) { //from https://gist.github.com/liangzan/807712/8fb16263cb39e8472d17aea760b6b1492c465af2
-    try { var files = fs.readdirSync(dirPath); }
-    catch(e) { return; }
-    if (files.length > 0)
-      for (var i = 0; i < files.length; i++) {
-        var filePath = dirPath + '/' + files[i];
-        if (fs.statSync(filePath).isFile())
-          fs.unlinkSync(filePath);
-        else
-          rmDir(filePath);
-      }
-    if (removeSelf)
-      fs.rmdirSync(dirPath);
-  };
+function rmDirTimed(dirPath, removeSelf, timeMin) { //from https://gist.github.com/liangzan/807712/8fb16263cb39e8472d17aea760b6b1492c465af2
+    try { files = fs.readdirSync(dirPath); }
+    catch(err) { return; }
+    if (files.length > 0) {
+        for (var i = 0; i < files.length; i++) {
+            var filePath = dirPath + '/' + files[i];
+            if (fs.statSync(filePath).isFile()) {
+                if (isOlderThan(filePath, timeMin)) {
+                    fs.unlinkSync(filePath);
+                }
+            }
+            else {
+                rmDirTimed(filePath, true, timeMin);
+            }
+        }
+    }
+    if (removeSelf) {
+        try { fs.rmdirSync(dirPath); } //only removes directories if they are empty
+        catch(err) { return; }
+    }
+}
+function isOlderThan(filePath, timeMin) {
+    const stat = fs.statSync(filePath);
+    const now = new Date().getTime();
+    const endTime = new Date(stat.ctime).getTime() + timeMin * 60 * 1000;
+    if (endTime < now) {
+        return true;
+    } else {
+        return false;
+    }
+}
 
   module.exports = {initBuckets};
